fix compile warning

+++ SequoiaDB/engine/pmd/pmdEnv.cpp
@@ -173,7 +173,7 @@ namespace engine
 
             /// Not the main thread
             if ( SIGFPE == sigNum &&
-                 ossGetCurrentProcessID() != ossGetCurrentThreadID() )
+                 ossGetCurrentProcessID() != (UINT32)ossGetCurrentThreadID() )
             {
                /// sleep 10 seconds
                ossSleep( 10 * OSS_ONE_SEC ) ;
